Move arrow utils to new crate re_arrow_util
