+++ crates/store/re_chunk/src/shuffle.rs
@@ -279,7 +279,7 @@ impl Chunk {
                     ArrowOffsets::try_from_lengths(sorted_arrays.iter().map(|array| array.len()))
                         .unwrap();
                 #[allow(clippy::unwrap_used)] // these are slices of the same outer array
-                let values = crate::arrow2_util::concat_arrays(&sorted_arrays).unwrap();
+                let values = re_arrow_util::arrow2_util::concat_arrays(&sorted_arrays).unwrap();
                 let validity = original
                     .validity()
                     .map(|validity| swaps.iter().map(|&from| validity.get_bit(from)).collect());
