+++ crates/store/re_chunk/src/slice.rs
@@ -5,10 +5,12 @@ use arrow2::array::{
 use itertools::Itertools;
 use nohash_hasher::IntSet;
 
+use re_arrow_util::arrow2_util;
+use re_arrow_util::arrow_util;
 use re_log_types::Timeline;
 use re_types_core::{ComponentDescriptor, ComponentName};
 
-use crate::{arrow2_util, arrow_util, Chunk, RowId, TimeColumn};
+use crate::{Chunk, RowId, TimeColumn};
 
 // ---
