+++ crates/store/re_chunk_store/Cargo.toml
@@ -27,6 +27,7 @@ deadlock_detection = ["parking_lot/deadlock_detection"]
 
 [dependencies]
 # Rerun dependencies:
+re_arrow_util.workspace = true
 re_byte_size.workspace = true
 re_chunk.workspace = true
 re_format.workspace = true
