+++ crates/store/re_dataframe/Cargo.toml
@@ -26,6 +26,7 @@ default = []
 
 [dependencies]
 # Rerun dependencies:
+re_arrow_util.workspace = true
 re_chunk.workspace = true
 re_chunk_store.workspace = true
 re_log.workspace = true
