+++ crates/store/re_dataframe/src/lib.rs
@@ -9,7 +9,7 @@ pub use self::query::QueryHandle;
 #[doc(no_inline)]
 pub use self::external::arrow2::chunk::Chunk as Arrow2Chunk;
 #[doc(no_inline)]
-pub use self::external::re_chunk::{arrow2_util::concatenate_record_batches, TransportChunk};
+pub use self::external::re_chunk::{concat_record_batches::concatenate_record_batches, TransportChunk};
 #[doc(no_inline)]
 pub use self::external::re_chunk_store::{
     ChunkStoreConfig, ChunkStoreHandle, ColumnSelector, ComponentColumnSelector, Index, IndexRange,
