+++ crates/store/re_grpc_client/Cargo.toml
@@ -20,6 +20,7 @@ all-features = true
 
 
 [dependencies]
+re_arrow_util.workspace = true
 re_chunk.workspace = true
 re_error.workspace = true
 re_log.workspace = true
