+++ crates/utils/re_arrow_util/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "re_arrow_util"
+authors.workspace = true
+description = "Helpers for working with arrow."
+edition.workspace = true
+homepage.workspace = true
+include.workspace = true
+license.workspace = true
+publish = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[lints]
+workspace = true
+
+[package.metadata.docs.rs]
+all-features = true
+
+
+[features]
+
+
+[dependencies]
+re_log.workspace = true
+re_tracing.workspace = true
+
+arrow.workspace = true
+arrow2.workspace = true
+itertools.workspace = true
