+++ crates/utils/re_arrow_util/README.md
@@ -1,0 +1,10 @@
+# re_arrow_util
+
+Part of the [`rerun`](https://github.com/rerun-io/rerun) family of crates.
+
+[![Latest version](https://img.shields.io/crates/v/re_arrow_util.svg)](https://crates.io/crates/re_arrow_util?speculative-link)
+[![Documentation](https://docs.rs/re_arrow_util/badge.svg)](https://docs.rs/re_arrow_util?speculative-link)
+![MIT](https://img.shields.io/badge/license-MIT-blue.svg)
+![Apache](https://img.shields.io/badge/license-Apache-blue.svg)
+
+Helpers for working with arrow
