+++ crates/utils/re_arrow_util/src/arrow2_util.rs
@@ -1,4 +1,3 @@
-use arrow::datatypes::Schema as ArrowSchema;
 use arrow2::{
     array::{
         Array as Arrow2Array, BooleanArray as Arrow2BooleanArray,
@@ -9,9 +8,7 @@ use arrow2::{
     datatypes::DataType as Arrow2Datatype,
     offset::Offsets as ArrowOffsets,
 };
-use itertools::Itertools;
-
-use crate::TransportChunk;
+use itertools::Itertools as _;
 
 /// Downcast an arrow array to another array, without having to go via `Any`.
 ///
@@ -453,44 +450,4 @@ pub fn take_array<A: Arrow2Array + Clone, O: arrow2::types::Index>(
         // Unwrap: that's initial type that we got.
         .unwrap()
         .clone()
-}
-
-// ---
-
-use arrow2::chunk::Chunk as Arrow2Chunk;
-
-/// Concatenate multiple [`TransportChunk`]s into one.
-///
-/// This is a temporary method that we use while waiting to migrate towards `arrow-rs`.
-/// * `arrow2` doesn't have a `RecordBatch` type, therefore we emulate that using our `TransportChunk`s.
-/// * `arrow-rs` does have one, and it natively supports concatenation.
-pub fn concatenate_record_batches(
-    schema: impl Into<ArrowSchema>,
-    batches: &[TransportChunk],
-) -> anyhow::Result<TransportChunk> {
-    let schema: ArrowSchema = schema.into();
-    anyhow::ensure!(
-        batches
-            .iter()
-            .all(|batch| batch.schema_ref().as_ref() == &schema),
-        "concatenate_record_batches: all batches must have the same schema"
-    );
-
-    let mut output_columns = Vec::new();
-
-    if !batches.is_empty() {
-        for (i, _field) in schema.fields.iter().enumerate() {
-            let arrays: Option<Vec<_>> = batches.iter().map(|batch| batch.column(i)).collect();
-            let arrays = arrays.ok_or_else(|| {
-                anyhow::anyhow!("concatenate_record_batches: all batches must have the same schema")
-            })?;
-            let array = concat_arrays(&arrays)?;
-            output_columns.push(array);
-        }
-    }
-
-    Ok(TransportChunk::new(
-        schema,
-        Arrow2Chunk::new(output_columns),
-    ))
 }
