+++ crates/utils/re_arrow_util/src/arrow_util.rs
@@ -3,7 +3,7 @@ use arrow::{
     buffer::{NullBuffer, OffsetBuffer},
     datatypes::{DataType, Field},
 };
-use itertools::Itertools;
+use itertools::Itertools as _;
 
 // ---
