+++ crates/utils/re_arrow_util/src/lib.rs
@@ -1,0 +1,7 @@
+//! Helpers for working with arrow
+
+pub mod arrow2_util;
+pub mod arrow_util;
+
+pub use arrow2_util::Arrow2ArrayDowncastRef;
+pub use arrow_util::ArrowArrayDowncastRef;
